src/cloud_functions/ingestion/api_ingestion/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-secret-manager
